Added initial plant food spawner to GameController.
This is an initial version of a 3-wave plant food spawner.
Each grid square is 100x100.
The first wave will do the squares in the ring around the origin.
The second wave will do the squares in the ring 1-square out from the origin.
The third wave will do the squares in the ring 2-squares out from the origin.
There is a 1 second pause before it starts planting food.
There is a 1 second pause between each item spawning.
There is a 1 second pause between waves.
Known issue: it doesn't track where it has spawned, so it will spawn multiple items in the same square. Once this is fixed, the expected result would be to have 3 rings of food when it was complete.