Fixed the issue with the game hanging. Also fixed the plant view being behind the background.
The food spawner was stuck in an infinite loop because of a logic issue.
Set the PlantFood prefab to a sortingOrder of 1 so that it can be seen.
Tweaked the food spawn settings so you could debug faster. There is a 10 second wait on start so you can move the snake out of the way of the spawning food.