Updated enemy movement after enemies spawn
Enemies now move in either vertical or hoizontal lines based on their spawn location. This is to keep the code simple.
Enemies lifetimes are set to expire as they get to the other side of the board.
Also, I upped the EnemyPoolLimit to 10 as the last wave (9) should be able to spawn up to 9 enemies at once.